+++ bundle-analyzer/src/bundles/ricos-editor-no-modal.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { RicosEditorType } from 'ricos-editor';
+import { RicosEditor } from 'ricos-editor';
 
 const modalSettings = {
   openModal: () => {},
@@ -7,5 +7,5 @@ const modalSettings = {
 };
 
 export default () => {
-  <RicosEditorType modalSettings={modalSettings} />;
+  <RicosEditor modalSettings={modalSettings} />;
 };