+++ packages/ricos-editor/web/src/index.ts
@@ -1,4 +1,4 @@
 export * from 'ricos-common';
 
 export { default as RicosEditor } from './RicosEditorSwitcher';
-export { RicosEditor as RicosEditorType } from './RicosEditor';
+export { RicosEditorRef as RicosEditorType } from './RicosEditorRef';